backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Talisman==0.8.1
python-dotenv==0.19.0
celery==5.1.2
redis==3.5.3
gitpython==3.1.18
docker==5.0.0
gunicorn==20.1.0
pytest==6.2.4
flask-swagger-ui==3.36.0
flask-limiter==1.4
flask-jwt-extended==4.4.4